moved files to correct folders and added template views
