+++ moesland-app/src/components/NewsItemContentView.js
