+++ moesland-app/src/components/ToolbarView.js
@@ -3,11 +3,10 @@ import { Text } from 'react-native';
 import { Header, Icon } from 'react-native-elements';
 import { useNavigation } from '@react-navigation/native';
 import { MOESLAND_GREEN, COLOR_WHITE } from '../constants/colors';
-import APP_TITLE from '../constants/strings'
+import { APP_TITLE } from '../constants/strings'
 
-const ToolbarView = ({ showBackButton, onPressMenu }) => {
+const ToolbarView = ({ showBackButton }) => {
   const navigation = useNavigation();
-
   const backButton = (
     <Icon
       name="arrow-back"
@@ -20,7 +19,7 @@ const ToolbarView = ({ showBackButton, onPressMenu }) => {
     <Header
       backgroundColor={MOESLAND_GREEN}
       leftComponent={showBackButton ? backButton : null}
-      centerComponent={<Text style={{ color: "#000000", fontSize: 30, fontWeight: 'bold' }}>CV Moesland</Text>}
+      centerComponent={<Text style={{ color: "#000000", fontSize: 30, fontWeight: 'bold' }}>{APP_TITLE}</Text>}
     />
   );
 };
