+++ moesland-app/src/routes/CalendarView.js
@@ -1,0 +1,10 @@
+import React, { Component } from 'react';
+import { Text } from 'react-native';
+
+export default class NewsItemListView extends Component {
+    render() {
+        return (
+            <Text>Kalender</Text>
+        )
+    }
+}
